requirements.txt
# Legal Risk Analysis System Dependencies

# Anthropic SDK for Claude API
anthropic>=0.40.0

# Document generation
python-docx>=1.1.0

# Data handling
pydantic>=2.0.0

# Web requests (for web research tools)
requests>=2.31.0
httpx>=0.27.0

# HTML/content processing
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Utilities
python-dateutil>=2.8.0

# Optional: For advanced features
# langchain>=0.1.0
# langgraph>=0.0.20
